+++ blackbox.php
@@ -10,9 +10,9 @@ use Innmind\BlackBox\{
 };
 
 Application::new($argv)
-    ->when(
-        \getenv('ENABLE_COVERAGE') !== false,
-        static fn(Application $app) => $app
+    ->map(static fn($app) => match (\getenv('BLACKBOX_ENV')) {
+        'extensive' => $app->scenariiPerProof(1_000),
+        'coverage' => $app
             ->codeCoverage(
                 CodeCoverage::of(
                     __DIR__.'/src/',
@@ -23,6 +23,7 @@ Application::new($argv)
                     ->enableWhen(true),
             )
             ->scenariiPerProof(50),
-    )
+        default => $app,
+    })
     ->tryToProve(Load::everythingIn(__DIR__.'/proofs/'))
     ->exit();