+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 6.1.0 - 2026-04-05
+
+### Added
+
+- `Innmind\Immutable\Validation::attempt()`
+- `Innmind\Immutable\Attempt::mapErrorCase()`
+- `Innmind\Immutable\Attempt::recoverCase()`
+- `Innmind\Immutable\Sequence::union()`
+- `Innmind\Immutable\Sequence\Union\Left`
+- `Innmind\Immutable\Sequence\Union\Right`
+- `Innmind\Immutable\Sequence\Union\Both`
+
 ## 6.0.0 - 2026-01-11
 
 ### Added